For cash register systems
This validator is intended to aid in implementing digital signatures in cash register systems according to the document Requirements and guidelines for implementing digital signatures in Cash Register Systems - http://skatteetaten.no/globalassets/standardformat-regnskapsaf-t/requirements-and-guidelines-for-implementing-digital-signatures-in-cash-register-systems.pdf

The signing methods to be used are RSA-SHA1-1024 or HMAC-SHA1-128. The validator generates a signature based on chosen signing method, data to be signed and key(s). This signature can then be verified with the signature generated in the cash register system.