Fictitious company Tøyen Toy Factory
The example file is produced by the Norwegian Tax Administration. All names and companies in the file are purely fictional. The example file is developed for testing and educational purposes, and does not represent the legal demands with respect to the bookkeeping regulations and/or the SAF-T Financial format.

For information on the legal demands, please see the Directorate of Taxes establishment of the SAF-T Financial format and the documentation: 
https://www.skatteetaten.no/rettskilder/type/kunngjoringer/standard-for-innhold-og-format-ved-gjengivelse-av-elektronisk-bokforte-opplysninger/
https://www.skatteetaten.no/bedrift-og-organisasjon/starte-og-drive/rutiner-regnskap-og-kassasystem/saf-t-regnskap/dokumentasjon/